+++ src/commonMain/kotlin/dev/inmo/krontab/StringParser.kt
@@ -14,14 +14,15 @@ typealias KrontabTemplate = String
 /**
  * Parse [incoming] string and adapt according to next format: "* * * * *" where order of things:
  *
- * * seconds
- * * minutes
- * * hours
- * * dayOfMonth
- * * month
- * * (optional) year
- * * (optional) (can be placed anywhere after month) (must be marked with `o` at the end, for example: 60o == +01:00) offset
- * * (optional) (can be placed anywhere after month) dayOfWeek
+ * * **seconds**
+ * * **minutes**
+ * * **hours**
+ * * **dayOfMonth**
+ * * **month**
+ * * **year** (optional)
+ * * **offset** (optional) (can be placed anywhere after month) (must be marked with `o` at the end, for example: 60o == +01:00)
+ * * **dayOfWeek** (optional) (can be placed anywhere after month)
+ * * **milliseconds** (optional) (can be placed anywhere after month) (must be marked with `ms` at the end, for example: 500ms; 100-200ms)
  *
  * And each one (except of offsets) have next format:
  *
@@ -48,18 +49,21 @@ typealias KrontabTemplate = String
  * * Months ranges can be found in [monthRange]
  * * Years ranges can be found in [yearRange] (in fact - any [Int])
  * * WeekDay (timezone) ranges can be found in [dayOfWeekRange]
+ * * Milliseconds ranges can be found in [millisecondsRange]
  *
  * Examples:
  *
  * * "0/5 * * * *" for every five seconds triggering
  * * "0/5,L * * * *" for every five seconds triggering and on 59 second
  * * "0/15 30 * * *" for every 15th seconds in a half of each hour
+ * * "0/15 30 * * * 500ms" for every 15th seconds in a half of each hour when milliseconds equal to 500
  * * "1 2 3 F,4,L 5" for triggering in near first second of second minute of third hour of fourth day of may
  * * "1 2 3 F,4,L 5 60o" for triggering in near first second of second minute of third hour of fourth day of may with timezone UTC+01:00
  * * "1 2 3 F,4,L 5 60o 0-2w" for triggering in near first second of second minute of third hour of fourth day of may in case if it will be in Sunday-Tuesday week days with timezone UTC+01:00
  * * "1 2 3 F,4,L 5 2021" for triggering in near first second of second minute of third hour of fourth day of may of 2021st year
  * * "1 2 3 F,4,L 5 2021 60o" for triggering in near first second of second minute of third hour of fourth day of may of 2021st year with timezone UTC+01:00
  * * "1 2 3 F,4,L 5 2021 60o 0-2w" for triggering in near first second of second minute of third hour of fourth day of may of 2021st year if it will be in Sunday-Tuesday week days with timezone UTC+01:00
+ * * "1 2 3 F,4,L 5 2021 60o 0-2w 500ms" for triggering in near first second of second minute of third hour of fourth day of may of 2021st year if it will be in Sunday-Tuesday week days with timezone UTC+01:00 when milliseconds will be equal to 500
  *
  * @return In case when offset parameter is absent in [incoming] will be used [createSimpleScheduler] method and
  * returned [CronDateTimeScheduler]. In case when offset parameter there is in [incoming] [KrontabTemplate] will be used
@@ -73,18 +77,22 @@ fun createSimpleScheduler(
     var offsetParsed: Int? = null
     var dayOfWeekParsed: Array<Byte>? = null
     var yearParsed: Array<Int>? = null
+    var millisecondsParsed: Array<Short>? = null
     val (secondsSource, minutesSource, hoursSource, dayOfMonthSource, monthSource) = incoming.split(" ").also {
         listOfNotNull(
             it.getOrNull(5),
             it.getOrNull(6),
-            it.getOrNull(7)
+            it.getOrNull(7),
+            it.getOrNull(8)
         ).forEach {
             val offsetFromString = parseOffset(it)
             val dayOfWeekFromString = parseWeekDay(it)
+            val millisecondsFromString = parseMilliseconds(it)
             offsetParsed = offsetParsed ?: offsetFromString
             dayOfWeekParsed = dayOfWeekParsed ?: dayOfWeekFromString
+            millisecondsParsed = millisecondsParsed ?: millisecondsFromString
             when {
-                dayOfWeekFromString != null || offsetFromString != null -> return@forEach
+                dayOfWeekFromString != null || offsetFromString != null || millisecondsFromString != null -> return@forEach
                 yearParsed == null -> {
                     yearParsed = parseYears(it)
                 }
@@ -100,10 +108,25 @@ fun createSimpleScheduler(
 
     return offsetParsed ?.let { offset ->
         createKronSchedulerWithOffset(
-            secondsParsed, minutesParsed, hoursParsed, dayOfMonthParsed, monthParsed, yearParsed, dayOfWeekParsed, TimezoneOffset(offset.minutes)
+            secondsParsed,
+            minutesParsed,
+            hoursParsed,
+            dayOfMonthParsed,
+            monthParsed,
+            yearParsed,
+            dayOfWeekParsed,
+            TimezoneOffset(offset.minutes),
+            millisecondsParsed ?: millisecondsArrayDefault
         )
     } ?: createKronScheduler(
-        secondsParsed, minutesParsed, hoursParsed, dayOfMonthParsed, monthParsed, yearParsed, dayOfWeekParsed
+        secondsParsed,
+        minutesParsed,
+        hoursParsed,
+        dayOfMonthParsed,
+        monthParsed,
+        yearParsed,
+        dayOfWeekParsed,
+        millisecondsParsed ?: millisecondsArrayDefault
     )
 }