+++ src/commonMain/kotlin/dev/inmo/krontab/internal/Defaults.kt
@@ -1,0 +1,4 @@
+package dev.inmo.krontab.internal
+
+internal const val millisecondsDefault: Short = 0
+internal val millisecondsArrayDefault: Array<Short> = arrayOf(millisecondsDefault)