+++ src/commonMain/kotlin/dev/inmo/krontab/internal/Ranges.kt
@@ -7,3 +7,4 @@ internal val dayOfMonthRange = 0 .. 30
 internal val hoursRange = 0 .. 23
 internal val minutesRange = 0 .. 59
 internal val secondsRange = minutesRange
+internal val millisecondsRange = 0 .. 999